Fix jitter caused by including headers/footers
It appears that something strange happens when a section has a header
and footer. It is not obvious but if a user scrolls to the bottom of a
QuickTableViewController, selects a radio button and then tries to
scroll up, there is a bit of jitter. This only happens when there are
headers/footers and the amount it shifts is completely dependent on the
header/footer height.